Apply target authorization to logging configuration
Move target authorization into a shared internal policy used by both
instantiate and logging configuration.

Apply the application target whitelist to logging handlers, formatters,
filters, queues, listeners, and external values. Preserve the legacy
blocklist as a deprecated fallback, mediate callable results, and reject
custom dictConfigClass implementations that would bypass authorization.

Propagate the policy through Hydra launches and distributed launchers, add
focused security and compatibility coverage, and document the shared
instantiate and logging migration path.
