+++ hydra/_internal/hydra.py
@@ -17,6 +17,7 @@ from omegaconf import (
 )
 from omegaconf.errors import InterpolationToMissingValueError
 
+from hydra._internal.target_policy import _get_active_target_whitelist
 from hydra._internal.utils import get_column_widths, run_and_report
 from hydra.core.config_loader import ConfigLoader
 from hydra.core.config_search_path import ConfigSearchPath
@@ -162,7 +163,9 @@ class Hydra:
         try:
             ret = run_job(
                 hydra_context=HydraContext(
-                    config_loader=self.config_loader, callbacks=callbacks
+                    config_loader=self.config_loader,
+                    callbacks=callbacks,
+                    target_whitelist=_get_active_target_whitelist(),
                 ),
                 task_function=task_function,
                 config=cfg,
@@ -211,7 +214,9 @@ class Hydra:
         sweeper = Plugins.instance().instantiate_sweeper(
             config=cfg,
             hydra_context=HydraContext(
-                config_loader=self.config_loader, callbacks=callbacks
+                config_loader=self.config_loader,
+                callbacks=callbacks,
+                target_whitelist=_get_active_target_whitelist(),
             ),
             task_function=task_function,
         )
