+++ plugins/hydra_joblib_launcher/hydra_plugins/hydra_joblib_launcher/_core.py
@@ -153,7 +153,11 @@ def launch(
     assert launcher.task_function is not None
     assert launcher.hydra_context is not None
 
-    configure_log(launcher.config.hydra.hydra_logging, launcher.config.hydra.verbose)
+    configure_log(
+        launcher.config.hydra.hydra_logging,
+        launcher.config.hydra.verbose,
+        target_whitelist=getattr(launcher.hydra_context, "target_whitelist", None),
+    )
     sweep_dir = Path(str(launcher.config.hydra.sweep.dir))
     sweep_dir.mkdir(parents=True, exist_ok=True)
