+++ plugins/hydra_ray_launcher/hydra_plugins/hydra_ray_launcher/_core_aws.py
@@ -76,7 +76,17 @@ def launch(
     )
     ray_cluster["setup_commands"] = setup_commands
 
-    configure_log(launcher.config.hydra.hydra_logging, launcher.config.hydra.verbose)
+    target_whitelist = getattr(launcher.hydra_context, "target_whitelist", None)
+    if target_whitelist is None:
+        configure_log(
+            launcher.config.hydra.hydra_logging, launcher.config.hydra.verbose
+        )
+    else:
+        configure_log(
+            launcher.config.hydra.hydra_logging,
+            launcher.config.hydra.verbose,
+            target_whitelist=target_whitelist,
+        )
     logging_config = cast(
         Dict[str, Any],
         OmegaConf.to_container(launcher.logging, resolve=True, enum_to_str=True),
