+++ tests/test_core_utils.py
@@ -1,14 +1,15 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 import sys
-from typing import Any
+from typing import Any, cast
 
 from omegaconf import OmegaConf, open_dict
 
 from hydra._internal.config_loader_impl import ConfigLoaderImpl
+from hydra._internal.target_policy import _get_active_target_whitelist
 from hydra._internal.utils import create_config_search_path
 from hydra.core import utils
 from hydra.core.hydra_config import HydraConfig
-from hydra.types import RunMode
+from hydra.types import HydraContext, RunMode
 
 
 def test_accessing_hydra_config(hydra_restore_singletons: Any) -> Any:
@@ -34,3 +35,30 @@ def test_py_version_resolver(hydra_restore_singletons: Any, monkeypatch: Any) ->
     assert OmegaConf.create({"key": "${python_version:major}"}).key == "3"
     assert OmegaConf.create({"key": "${python_version:minor}"}).key == "3.8"
     assert OmegaConf.create({"key": "${python_version:micro}"}).key == "3.8.2"
+
+
+def test_run_job_reestablishes_target_whitelist(monkeypatch: Any) -> None:
+    expected = ("tests.test_core_utils.Allowed",)
+    sentinel = object()
+
+    def fake_run_job(**kwargs: Any) -> object:
+        assert _get_active_target_whitelist() == expected
+        return sentinel
+
+    monkeypatch.setattr(utils, "_run_job", fake_run_job)
+    hydra_context = HydraContext(
+        config_loader=cast(Any, object()),
+        callbacks=cast(Any, object()),
+        target_whitelist=expected,
+    )
+
+    result = utils.run_job(
+        task_function=cast(Any, object()),
+        config=OmegaConf.create(),
+        job_dir_key="unused",
+        job_subdir_key=None,
+        hydra_context=hydra_context,
+    )
+
+    assert result is sentinel
+    assert _get_active_target_whitelist() is None
